fix MCP tool registration with GAITK, all tests passed
