+++ examples/agentcore/agent/mcp_tool_manager.py
@@ -1,14 +1,58 @@
 #!/usr/bin/env python3
 """MCP Tool Manager for integrating MCP servers with Generative AI Toolkit agents."""
 
+import asyncio
 import logging
 import os
 
 from simple_mcp_client import SimpleMcpClient
 
 from generative_ai_toolkit.agent import BedrockConverseAgent
+from generative_ai_toolkit.agent.tool import Tool
 
 logger = logging.getLogger(__name__)
+
+
+class McpTool(Tool):
+    """A Tool implementation that wraps MCP tool calls."""
+
+    def __init__(self, mcp_tool, tool_manager: "McpToolManager"):
+        self.mcp_tool = mcp_tool
+        self.tool_manager = tool_manager
+        self._tool_spec = {
+            "name": mcp_tool.name,
+            "description": mcp_tool.description,
+            "inputSchema": {"json": mcp_tool.inputSchema},
+        }
+
+    @property
+    def tool_spec(self):
+        return self._tool_spec
+
+    def invoke(self, **kwargs):
+        """Invoke the MCP tool synchronously."""
+        try:
+            # Run the async operation synchronously to match Tool protocol
+            return asyncio.run(self._async_invoke(**kwargs))
+        except Exception as e:
+            logger.error(f"Error calling MCP tool {self.mcp_tool.name}: {e}")
+            return f"Error calling {self.mcp_tool.name}: {str(e)}"
+
+    async def _async_invoke(self, **kwargs):
+        """Internal async implementation of MCP tool invocation."""
+        # Create a fresh client for each tool call to avoid event loop issues
+        client = await self.tool_manager.get_mcp_client()
+
+        result = await client.call_tool(self.mcp_tool.name, kwargs)
+
+        # Extract text content from MCP result
+        if hasattr(result, "content") and result.content:
+            if hasattr(result.content[0], "text"):
+                return result.content[0].text
+            else:
+                return str(result.content[0])
+        else:
+            return str(result)
 
 
 class McpToolManager:
@@ -57,45 +101,11 @@ class McpToolManager:
             for mcp_tool in tools_result.tools:
                 logger.info(f"Registering tool: {mcp_tool.name}")
 
-                # Create a wrapper function for the MCP tool
-                def create_tool_wrapper(tool_name: str):
-                    async def tool_wrapper(**kwargs) -> str:
-                        """Wrapper function to call MCP tool."""
-                        try:
-                            # Create a fresh client for each tool call to avoid event loop issues
-                            client = await self.get_mcp_client()
-
-                            result = await client.call_tool(tool_name, kwargs)
-
-                            # Extract text content from MCP result
-                            if hasattr(result, "content") and result.content:
-                                if hasattr(result.content[0], "text"):
-                                    return result.content[0].text
-                                else:
-                                    return str(result.content[0])
-                            else:
-                                return str(result)
-
-                        except Exception as e:
-                            logger.error(f"Error calling MCP tool {tool_name}: {e}")
-                            return f"Error calling {tool_name}: {str(e)}"
-
-                    return tool_wrapper
-
-                # Create the wrapper and register it
-                tool_func = create_tool_wrapper(mcp_tool.name)
-                tool_func.__name__ = mcp_tool.name
-                tool_func.__doc__ = mcp_tool.description
-
-                # Create tool_spec in Bedrock format from MCP tool schema
-                tool_spec = {
-                    "name": mcp_tool.name,
-                    "description": mcp_tool.description,
-                    "inputSchema": {"json": mcp_tool.inputSchema},
-                }
+                # Create MCP tool wrapper
+                tool = McpTool(mcp_tool, self)
 
                 # Register with the agent
-                agent.register_tool(tool_func, tool_spec=tool_spec)
+                agent.register_tool(tool)
 
             self.tools_registered = True
             logger.info("MCP tools registered successfully")
