+++ examples/agentcore/mcp-server/mcp_server.py
@@ -1,10 +1,11 @@
 #!/usr/bin/env python3
-"""MCP Server for AgentCore Runtime using Pydantic models with FastMCP."""
+"""MCP Server for AgentCore Runtime using modular weather tools with FastMCP."""
 
 import logging
 
+from get_forecast_tool import register_get_forecast_tool
+from get_weather_tool import register_get_weather_tool
 from mcp.server.fastmcp import FastMCP
-from weather_tools import register_weather_tools
 
 # Configure logging
 logging.basicConfig(level=logging.INFO, format="%(name)s - %(levelname)s - %(message)s")
@@ -15,8 +16,9 @@ logger.info("MCP Server starting")
 # Create FastMCP server with correct configuration for AgentCore Runtime
 mcp = FastMCP(host="0.0.0.0", stateless_http=True)
 
-# Register all weather tools
-register_weather_tools(mcp)
+# Register weather tools from separate modules
+register_get_weather_tool(mcp)
+register_get_forecast_tool(mcp)
 
 if __name__ == "__main__":
     logger.info("Starting Weather Tools MCP Server with FastMCP")
